Move the function+isa generation into cranelfit-fuzzgen
